# perf(813): Avoid materialization of product values in subscriptions