# Description of Changes

- Fixes https://github.com/clockworklabs/SpacetimeDB/issues/813.

```
Benchmarking full-scan: Collecting 100 samples in estimated 
full-scan               time:   [162.84 ms 164.37 ms 166.88 ms]
                        change: [-47.765% -47.089% -46.260%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking query-indexes-multi: Collecting 100 samples in 
query-indexes-multi     time:   [1.1162 µs 1.1202 µs 1.1274 µs]
                        change: [-6.4601% -5.3782% -4.0002%] (p = 0.00 < 0.05)
                        Performance has improved.
```

# API and ABI breaking changes

None

# Expected complexity level and risk

2