@joshua-spacetime 2024-03-12
Here are the numbers I get compared to latest master:
```
full-scan               time:   [82.200 ms 82.378 ms 82.581 ms]
                        change: [-16.491% -16.056% -15.598%] (p = 0.00 < 0.05)
                        Performance has improved.

full-join               time:   [236.58 µs 236.81 µs 237.07 µs]
                        change: [-8.2742% -8.0187% -7.7901%] (p = 0.00 < 0.05)
                        Performance has improved.

incr-select             time:   [666.93 ns 667.28 ns 667.64 ns]
                        change: [+53.585% +53.879% +54.170%] (p = 0.00 < 0.05)
                        Performance has regressed.

incr-join               time:   [3.6968 µs 3.7051 µs 3.7123 µs]
                        change: [-18.685% -18.431% -18.173%] (p = 0.00 < 0.05)
                        Performance has improved.
```

I haven't reviewed this patch yet, so I don't know where the regression in `incr-select` is coming from. I'll get back to you once I have more information. But pretty good numbers otherwise.