@jdetter 2024-03-13
Master benchmark on 14900k:
```
    Finished bench [optimized + debuginfo] target(s) in 0.15s
     Running benches/subscription.rs (/home/ubuntu/SpacetimeDB/target/release/deps/subscription-7c24f5921892496c)
Benchmarking full-scan: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 29.7s, or reduce sample count to 10.
full-scan               time:   [243.00 ms 243.98 ms 244.95 ms]

full-join               time:   [686.00 µs 686.16 µs 686.33 µs]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high severe

incr-select             time:   [481.07 ns 481.26 ns 481.45 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

incr-join               time:   [6.4075 µs 6.4135 µs 6.4199 µs]
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

query-indexes-multi     time:   [925.65 ns 925.95 ns 926.25 ns]
```

crntril/borrow-eq:
```
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 12.0s, or reduce sample count to 40.
full-scan               time:   [113.10 ms 113.30 ms 113.50 ms]
                        change: [-53.768% -53.563% -53.363%] (p = 0.00 < 0.05)
                        Performance has improved.

full-join               time:   [573.18 µs 573.49 µs 573.83 µs]
                        change: [-16.468% -16.422% -16.372%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild

incr-select             time:   [455.01 ns 455.35 ns 455.72 ns]
                        change: [-5.4337% -5.3479% -5.2623%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

incr-join               time:   [5.0606 µs 5.0649 µs 5.0692 µs]
                        change: [-21.097% -21.008% -20.923%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

query-indexes-multi     time:   [843.18 ns 843.48 ns 843.79 ns]
                        change: [-8.9223% -8.8640% -8.8035%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  7 (7.00%) high mild
  3 (3.00%) high severe
```