@cloutiertyler 2024-03-14
I, too, am a bit concerned about modifying the datastore trait in this way, if only because it may have implications on MVCC which I need to be aware of. Could you please make a post here (or better yet a comment in the codebase) which explains the purpose of this lifetime and why it is necessary?