@gefjon 2024-03-14
I'd really like for these two tickets to get separate PRs so we can measure the performance impact of the solutions separately.