@joshua-spacetime 2024-03-15
```
full-scan               time:   [84.373 ms 84.635 ms 84.909 ms]
                        change: [-16.825% -16.324% -15.837%] (p = 0.00 < 0.05)
                        Performance has improved.

full-join               time:   [241.59 µs 242.01 µs 242.56 µs]
                        change: [-13.306% -12.923% -12.537%] (p = 0.00 < 0.05)
                        Performance has improved.
```